requirement.txt
certifi==2017.11.5
chardet==3.0.4
Django==1.9.7
django-activity-stream==0.6.2
django-auth0==0.0.2
django-cors-headers==1.1.0
django-el-pagination==2.1.2
django-guardian==1.4.9
django-model-utils==2.5.1
django-notifications==0.1.dev0
django-notifications-hq==1.2
django-smuggler==0.7.0
django-widget-tweaks==1.4.1
djangorestframework==3.4.4
idna==2.6
jsonfield==1.0.3
notifications==0.3.2
olefile==0.44
Pillow==4.3.0
pkg-resources==0.0.0
psycopg2==2.7.3.2
pytz==2017.3
requests==2.11.0
six==1.11.0
urllib3==1.22
whitenoise==3.3.1
